// src/app/models/user.model.ts
export interface User {
  id:          number;
  firstName:  string;
  lastName:   string;
  email:       string;
  contactNumber:     string;
  gender:      string;
  designation: string;
  dob:         Date;
  age:         number;
  photo:       string;
  country:     string;
  state:       string;
  city:        string;
  pinCode:     number;
}